First rough implementation of onFollow
Implementation of onFollow by polling the API.

Issue realized: The API only allows for 30 requests per minute but for
each channel monitored, one request is required in addition to a few
more requests for converting between username and id. Sending a Bearer
token will increase the limit to 60.

Testing required: Using the chat token for the API to increase limit?

Missing: Implementation of the preparation in case a new Follow-Poll is
created -> Request the followers of the channel and store the most
recent one in map